+++ src/impls.rs
@@ -1638,8 +1638,10 @@ macro_rules! impl_f16 {
                 value: Self::NonAtomicType,
                 order: Ordering,
             ) -> Self::NonAtomicType {
-                self.fetch_update(Ordering::Relaxed, order, |x| Some(x + value))
-                    .unwrap()
+                self.fetch_update(order, crate::atomic::load_ordering(order), |x| {
+                    Some(x + value)
+                })
+                .unwrap()
             }
 
             #[inline(always)]
@@ -1648,8 +1650,10 @@ macro_rules! impl_f16 {
                 value: Self::NonAtomicType,
                 order: Ordering,
             ) -> Self::NonAtomicType {
-                self.fetch_update(Ordering::Relaxed, order, |x| Some(x - value))
-                    .unwrap()
+                self.fetch_update(order, crate::atomic::load_ordering(order), |x| {
+                    Some(x - value)
+                })
+                .unwrap()
             }
 
             #[inline(always)]
@@ -1658,7 +1662,7 @@ macro_rules! impl_f16 {
                 value: Self::NonAtomicType,
                 order: Ordering,
             ) -> Self::NonAtomicType {
-                self.fetch_update(Ordering::Relaxed, order, |x| {
+                self.fetch_update(order, crate::atomic::load_ordering(order), |x| {
                     Some(Self::NonAtomicType::min(x, value))
                 })
                 .unwrap()
@@ -1670,7 +1674,7 @@ macro_rules! impl_f16 {
                 value: Self::NonAtomicType,
                 order: Ordering,
             ) -> Self::NonAtomicType {
-                self.fetch_update(Ordering::Relaxed, order, |x| {
+                self.fetch_update(order, crate::atomic::load_ordering(order), |x| {
                     Some(Self::NonAtomicType::max(x, value))
                 })
                 .unwrap()
@@ -2106,7 +2110,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_recip(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).recip().to_bits())
                     })
                     .unwrap();
@@ -2114,7 +2118,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_to_degrees(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).to_degrees().to_bits())
                     })
                     .unwrap();
@@ -2122,7 +2126,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_to_radians(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).to_radians().to_bits())
                     })
                     .unwrap();
@@ -2131,7 +2135,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_div_euclid(&self, rhs: Self::NonAtomicType, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).div_euclid(rhs).to_bits())
                     })
                     .unwrap();
@@ -2140,7 +2144,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_rem_euclid(&self, rhs: Self::NonAtomicType, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).rem_euclid(rhs).to_bits())
                     })
                     .unwrap();
@@ -2149,7 +2153,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_floor(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).floor().to_bits())
                     })
                     .unwrap();
@@ -2158,7 +2162,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_ceil(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).ceil().to_bits())
                     })
                     .unwrap();
@@ -2167,7 +2171,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_round(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).round().to_bits())
                     })
                     .unwrap();
@@ -2176,7 +2180,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_trunc(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).trunc().to_bits())
                     })
                     .unwrap();
@@ -2185,7 +2189,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_fract(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).fract().to_bits())
                     })
                     .unwrap();
@@ -2194,7 +2198,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_abs(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).abs().to_bits())
                     })
                     .unwrap();
@@ -2203,7 +2207,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_signum(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).signum().to_bits())
                     })
                     .unwrap();
@@ -2212,7 +2216,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_copysign(&self, sign: Self::NonAtomicType, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).copysign(sign).to_bits())
                     })
                     .unwrap();
@@ -2221,7 +2225,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_powi(&self, n: i32, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).powi(n).to_bits())
                     })
                     .unwrap();
@@ -2230,7 +2234,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_powf(&self, n: Self::NonAtomicType, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).powf(n).to_bits())
                     })
                     .unwrap();
@@ -2239,7 +2243,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_sqrt(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).sqrt().to_bits())
                     })
                     .unwrap();
@@ -2248,7 +2252,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_exp(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).exp().to_bits())
                     })
                     .unwrap();
@@ -2257,7 +2261,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_exp2(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).exp2().to_bits())
                     })
                     .unwrap();
@@ -2266,7 +2270,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_ln(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).ln().to_bits())
                     })
                     .unwrap();
@@ -2275,7 +2279,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_log(&self, base: Self::NonAtomicType, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).log(base).to_bits())
                     })
                     .unwrap();
@@ -2284,7 +2288,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_log2(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).log2().to_bits())
                     })
                     .unwrap();
@@ -2293,7 +2297,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_log10(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).log10().to_bits())
                     })
                     .unwrap();
@@ -2302,7 +2306,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_cbrt(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).cbrt().to_bits())
                     })
                     .unwrap();
@@ -2311,7 +2315,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_sin(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).sin().to_bits())
                     })
                     .unwrap();
@@ -2320,7 +2324,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_cos(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).cos().to_bits())
                     })
                     .unwrap();
@@ -2329,7 +2333,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_tan(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).tan().to_bits())
                     })
                     .unwrap();
@@ -2338,7 +2342,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_asin(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).asin().to_bits())
                     })
                     .unwrap();
@@ -2347,7 +2351,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_acos(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).acos().to_bits())
                     })
                     .unwrap();
@@ -2356,7 +2360,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_atan(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).atan().to_bits())
                     })
                     .unwrap();
@@ -2365,7 +2369,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_exp_m1(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).exp_m1().to_bits())
                     })
                     .unwrap();
@@ -2374,7 +2378,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_ln_1p(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).ln_1p().to_bits())
                     })
                     .unwrap();
@@ -2383,7 +2387,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_sinh(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).sinh().to_bits())
                     })
                     .unwrap();
@@ -2392,7 +2396,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_cosh(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).cosh().to_bits())
                     })
                     .unwrap();
@@ -2401,7 +2405,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_tanh(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).tanh().to_bits())
                     })
                     .unwrap();
@@ -2410,7 +2414,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_asinh(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).asinh().to_bits())
                     })
                     .unwrap();
@@ -2419,7 +2423,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_acosh(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).acosh().to_bits())
                     })
                     .unwrap();
@@ -2428,7 +2432,7 @@ macro_rules! impl_f16 {
             #[inline(always)]
             fn fetch_atanh(&self, order: Ordering) {
                 self.0
-                    .fetch_update(Ordering::Relaxed, order, |x| {
+                    .fetch_update(order, crate::atomic::load_ordering(order), |x| {
                         Some(Self::NonAtomicType::from_bits(x).atanh().to_bits())
                     })
                     .unwrap();